Improve map detection logic
Check svMap to see if it's an instance of google.maps.Map, rather than just
an object. Similarly, check localMap to see if it's an instance of L.map.

Unfortunately, Strava's pageView.mapContext().map() is not an instance of
L.map, but it's so specific that it seems reasonable to assume it's the
object we're looking for if it exists.

Signed-off-by: Jim Mattson <jsmattsonjr@gmail.com>